Refactored to live in a PSR-0 namespace and added autoload lines to composer.json
