+++ composer.json
@@ -1,5 +1,8 @@
 {
     "name": "aligent/aligent-io",
     "description": "Delimited and fixed width file reader.",
-    "require": {}
+    "require": {},
+    "autoload":{
+        "psr-0": {"" : "src/" }
+    }
 }
