+++ src/Aligent/IO/Exception/IOEndOfLineSeparatorContainedInLineException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when attempting to write line which contains an end of line separator character.
+ */
+class IOEndOfLineSeparatorContainedInLineException extends IOException {}
+
