+++ src/Aligent/IO/Exception/IOException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * A generic exception for the IO package 
+ */
+class IOException extends \Exception {}
+
