+++ src/Aligent/IO/Exception/IOFieldOverflowException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when a field exceeds its fieldWidth
+ */
+class IOFieldOverflowException extends IOException {}
