+++ src/Aligent/IO/Exception/IOIllegalParameterException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: kath.young
+ * Date: 9/19/17
+ * Time: 9:47 AM
+ */
+
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when an illegal parameter is passed to a function in the IO
+ * package.
+ */
+class IOIllegalParameterException extends IOException{
+
+}
