+++ src/Aligent/IO/Exception/IOInputTooLongException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when the input string is longer than the sum of fieldWidths
+ */
+class IOInputTooLongException extends IOException {}
