+++ src/Aligent/IO/Exception/IOInputTooShortException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when the sum of the fieldWidths exceeds the length of the input string
+ */
+class IOInputTooShortException extends IOException {}
