+++ src/Aligent/IO/Exception/IONotImplementedYetException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+* Thrown when attempting to use unimplemented functionality
+*/
+class IONotImplementedYetException extends IOException {}
+
