+++ src/Aligent/IO/Exception/IOObjectAlreadyInitializedException.php
@@ -1,0 +1,8 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when attempting to initialize an object that has already been initialized.
+ */
+class IOObjectAlreadyInitializedException extends IOException {}
+
