+++ src/Aligent/IO/Exception/IOTooFewFieldsException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when there are less fields than fieldWidths
+ */
+class IOTooFewFieldsException extends IOException {}
