+++ src/Aligent/IO/Exception/IOTooManyFieldsException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Aligent\IO\Exception;
+
+/**
+ * Thrown when there are more fields than fieldWidths
+ */
+class IOTooManyFieldsException extends IOException {}
