+++ src/Aligent/IO/IOCSVFileWriter.php
@@ -1,6 +1,7 @@
 <?php
+namespace Aligent\IO;
 
-require_once 'IOFileWriter.php';
+use Aligent\IO\IOFileWriter;
 
 /**
  * class IOCSVFileWriter
