+++ src/Aligent/IO/IOLineFileWriter.php
@@ -1,7 +1,5 @@
 <?php
-
-require_once 'IO.php';
-require_once 'IOFileWriter.php';
+namespace Aligent\IO;
 
 /**
  * @author luke
