+++ tests/IO/IOArrayIteratorTest.php
@@ -1,6 +1,5 @@
 <?php
-
-require_once dirname(__FILE__) . '/../../IO/IOArrayIterator.php';
+require_once dirname(__FILE__) . "/../../vendor/autoload.php";
 
 /**
  * Test class for IOArrayIterator.
